requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
redis[hiredis]>=5.0.0
celery[redis]>=5.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pandas>=2.2.0
pandas_ta>=0.3.14b
numpy>=1.26.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
